Expose all available data as telemetry
We now send all the available data in the system is a structured way.
We ignore the interface required by the backend guys since they
don't really know what they want, and we want to allow ourselves
the freedom to modify the interface later.

So right now we send everything to a designated REST endpoint, and if
required later - we can add proxy that will adapt the data to the
operational server.

Signed-off-by: Avi Shukron <avraham.shukron@gmail.com>